Update experty-tge.sol
This should be tested but the idea is to give an option to the token buyer to claim his tokens himself by sending a transaction with a 0 ETH value. It's not easy for most investors to call a function by themselves. An other solution is to have the team to call those functions but it creates inequalities (who is receiving first ? aka the Enigma ICO...). This way anyone can claim the tokens whenever he wants by just sending 0 ETH after the sale.